cambios 23-01 17:24

+++ src/main/java/com/utp/venta/Repository/LeadRepositoryFromWSP.java
@@ -8,12 +8,16 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
-public interface LeadRepositoryFromWSP extends CrudRepository<LeadFromWSP, Long> {
+public interface LeadRepositoryFromWSP extends CrudRepository<LeadFromWSP, Integer> {
     LeadFromWSP findByDni(String dni);
 
     @Query("SELECT lw FROM LeadFromWSP lw WHERE lw.lead_id = :lead")
     LeadFromWSP findByLead(@Param("lead") Lead lead);
+
+    //LeadFromWSP findByLead_id(Integer leadId);
+
 
 }
